PEAR/requirements.txt
torch==2.0.1+cu118
torchvision
torchaudio

# 其他 PyTorch 扩展库
torchgeometry
torchmetrics

open3d==0.17.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80

decord
cmake==4.0.2

ConfigArgParse==1.7.1
configer==1.3.1
configparser==7.2.0


gradio==4.44.1
gradio_client==1.3.0
huggingface-hub==0.25.1

imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2


kornia==0.8.1
kornia_rs==0.1.9

numpy==1.22.4
omegaconf==2.3.0

pillow==10.4.0

plyfile==1.1

pyarrow==20.0.0
pycolmap==0.4.0


pynvml==12.0.0
PyOpenGL==3.1.0
pyOpenSSL==25.3.0
pyparsing==3.2.3
pyquaternion==0.9.9
pyrender==0.1.45
pyvista==0.46.4
PyYAML==6.0.2

rich==14.0.0
roma==1.5.2.1

safetensors==0.5.3
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.13.1
scooby==0.11.0
seaborn==0.13.2
smplx==0.1.28

urllib3==2.5.0
utils3d==0.1.1
ultralytics
einops
timm
colored
fvcore
iopath
lightning
webdataset